move algorithm commands from main to aux_functions
