+++ src-tauri/src/aux_functions/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod commands;
 pub(crate) mod edge_factory;
 pub(crate) mod graph_init;
 pub(crate) mod take_snapshot;
