+++ src-tauri/src/main.rs
@@ -3,20 +3,11 @@ mod aux_data_structures;
 mod graph;
 // Reference: https://rfdonnelly.github.io/posts/tauri-async-rust-process/
 mod aux_functions;
-use algorithms::articulation_point::articulation_point;
-use algorithms::globalmincut::karger_stein_gmincut;
-use algorithms::stoer_wagner::stoer_wagner;
-use aux_data_structures::neighbor_info::{Neighbor, NeighborInfo};
-use aux_data_structures::stoer_wagner_ds::StoerWagnerGraph;
-use aux_functions::graph_init::load_graph;
-use petgraph::graphmap::Nodes;
-use petgraph::stable_graph::NodeIndex;
-
+use aux_functions::commands::{run_articulation_point, run_global_mincut, run_stoer_wagner};
 use tauri::Manager;
 use tokio::sync::{mpsc, Mutex};
 use tracing::info;
 use tracing_subscriber;
-
 struct AsyncProcInputTx {
     inner: Mutex<mpsc::Sender<String>>,
 }
@@ -86,43 +77,4 @@ async fn async_process_model(
     }
 
     Ok(())
-}
-
-#[tauri::command]
-fn run_articulation_point(nodes: Vec<NeighborInfo>) -> String {
-    // Assemble a vector of nodes and their neighbors
-    println!("{:?}", nodes);
-    let mut graph = load_graph(nodes.clone());
-    let articulation_points = articulation_point(graph.clone());
-    let mut output = String::new();
-    output.push_str("Output: ");
-    for pt in articulation_points {
-        println!("Articulation point: {:?}", pt);
-        output.push_str(&graph.g.node_weight(pt).unwrap().name);
-    }
-    output
-}
-
-#[tauri::command]
-fn run_global_mincut(nodes: Vec<NeighborInfo>) -> f64 {
-    let graph = load_graph(nodes.clone());
-    println!("Hit global mincut");
-    let order = graph.get_order();
-    let global_min_cut = karger_stein_gmincut(&graph.clone(), order as i32);
-    global_min_cut
-}
-
-#[tauri::command]
-fn run_stoer_wagner(nodes: Vec<NeighborInfo>) -> String {
-    let graph = load_graph(nodes.clone());
-    println!("Hit Stoer-Wagner");
-    let graph_sw = &mut StoerWagnerGraph::new(graph.clone());
-    let stoer_wagner = stoer_wagner(&mut graph_sw.clone());
-    let output = format!(
-        "Stoer-Wagner: {} {} {}",
-        stoer_wagner.get_a(),
-        stoer_wagner.get_b(),
-        stoer_wagner.get_weight()
-    );
-    output
 }
